+++ CHANGELOG.rst
@@ -6,6 +6,16 @@ dcicutils
 Change Log
 ----------
 
+
+1.21.0
+======
+
+**PR 149: Add misc_utils.json_leaf_subst**
+
+* Add ``misc_utils.json_leaf_subst`` to do substitutions at the leaves
+  (atomic parts) of a JSON object.
+
+
 1.20.0
 ======